SearchKit - Break part of Run action into a base class
